requirements.txt
streamlit
google.generativeai
PILLOW
